+++ CHANGELOG.md
@@ -324,3 +324,4 @@
 - 2024-03-09: Integ find email, task, icp, mutate, +6
 - 2024-03-10: Shutdown sender, contacts, follow up, ready pool, +2
 - 2024-03-10: Builder descend, find email, main, mod, +6
+- 2024-03-11: Features lead, leads, deal, contacts, +2